cleaned code with Clean Up

+++ src/main/java/com/existingeevee/moretcon/block/blocktypes/BlockFalling.java
@@ -41,7 +41,7 @@ public class BlockFalling extends BlockBase {
 
 			if (!fallInstantly && worldIn.isAreaLoaded(pos.add(-32, -32, -32), pos.add(32, 32, 32))) {
 				if (!worldIn.isRemote) {
-					EntityFallingBlock entityfallingblock = new EntityFallingBlock(worldIn, (double) pos.getX() + 0.5D, (double) pos.getY(), (double) pos.getZ() + 0.5D, worldIn.getBlockState(pos));
+					EntityFallingBlock entityfallingblock = new EntityFallingBlock(worldIn, pos.getX() + 0.5D, pos.getY(), pos.getZ() + 0.5D, worldIn.getBlockState(pos));
 					// this.onStartFalling(entityfallingblock);
 					worldIn.spawnEntity(entityfallingblock);
 				}
@@ -51,7 +51,7 @@ public class BlockFalling extends BlockBase {
 				BlockPos blockpos;
 
 				for (blockpos = pos.down(); (worldIn.isAirBlock(blockpos) || canFallThrough(worldIn.getBlockState(blockpos))) && blockpos.getY() > 0; blockpos = blockpos.down()) {
-					;
+
 				}
 
 				if (blockpos.getY() > 0) {
