+++ src/main/java/com/existingeevee/moretcon/traits/modifiers/Shocked.java
@@ -32,7 +32,7 @@ public class Shocked extends ProjectileModifierTrait {
 	public void afterHit(EntityProjectileBase projectile, World world, ItemStack ammoStack, EntityLivingBase attacker, Entity target, double impactSpeed) {
 		if (!world.isRemote) {
 			if (target instanceof EntityLivingBase) {
-				float damage = MathHelper.ceil((double) MathHelper.sqrt(projectile.motionX * projectile.motionX + projectile.motionY * projectile.motionY + projectile.motionZ * projectile.motionZ) * projectile.getDamage());
+				float damage = MathHelper.ceil(MathHelper.sqrt(projectile.motionX * projectile.motionX + projectile.motionY * projectile.motionY + projectile.motionZ * projectile.motionZ) * projectile.getDamage());
 				damage += (projectile.getIsCritical() ? random.nextInt((int) damage / 2 + 2) : 0);
 				world.spawnEntity(new EntityShock(world, projectile, (EntityLivingBase) target, damage, projectile.isWet() || projectile.isInWater() || world.isRainingAt(projectile.getPosition().up())));
 			}
